refactor: fix code quality issues — move tests to external files, fix doc comments and log messages
E2 review fixes:
- Remove unused FeishuParser import check (not present in chat.rs)
- Update DslParser warn message to be renderer-agnostic
- Fix ProcessorRegistry::new() doc comment (no duplicate FeishuMessageCleaner, add SessionRouter)
- Move feishu_parser inline tests to feishu_parser_tests.rs
- Move session_router inline tests to session_router_tests.rs
- Move outbound_raw_log inline tests to outbound_raw_log_tests.rs
- Add DSL skip rendering tests for TerminalRenderer (public API)

Source: user
Type: refactor
